darwin: introduce darwin na(t)ive snapshotter
Darwin snapshotter is a snapshotter implementation which works on
darwin platform.  It is backed with sparsebundle images (.sparsebundle)
which is created from each layer.  Due to lack of public interface (or
API) to create sparsebundle images on macOS, each image creation and
mount operation invoke hdiutil command, which is unfortunatelly costly
(in terms of both storage space required and time completing each layer
operation).

But with this snapshotter, mount package doesn't depend on bind mount
features, which is transparent from other mount implementations.

This is originally decoupled from containerd pull request (5935):
https://github.com/containerd/containerd/pull/5935/commits

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>